refactor: es descriptor
